separated app logic to modules
