+++ modules/input_validators.rb
@@ -1,0 +1,14 @@
+module InputValidators
+  def validate_input(input, max_val)
+    loop do
+      input_is_valid = input.is_a?(Integer) && input.between?(1, max_val)
+      break if input_is_valid
+
+      print 'Please enter a valid input: '
+
+      input = gets.chomp.to_i
+    end
+
+    input
+  end
+end
